fix: debugging seg fault
