+++ v2/Makefile
@@ -6,7 +6,7 @@
 #    By: maalexan <maalexan@student.42sp.org.br>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/01/22 11:13:39 by maalexan          #+#    #+#              #
-#    Updated: 2023/09/05 12:23:31 by maalexan         ###   ########.fr        #
+#    Updated: 2023/09/07 15:22:33 by maalexan         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -61,7 +61,8 @@ FUN			:=	main.c \
 				token_setter.c \
 				token_str.c \
 				token_utils.c \
-				token.c 
+				token.c \
+				t_printers.c
 
 # Object files
 OBJ			:=	$(FUN:%.c=$(OBJ_DIR)/%.o)
